backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
PyJWT>=2.8.0
bcrypt==4.1.1
pydantic[email]==2.5.0
httpx==0.25.2
python-multipart==0.0.6
